test(types): extract StepOutput tests to separate file
Move StepOutput type tests from dsl-types.test-d.ts to a new step-output.test-d.ts file, removing
the tests from the original file and creating a dedicated test file for StepOutput utility type
